Move fips from ui to wxui
